move tests to a separate module
